windows/requirements.txt
# Video Chopper - Windows Dependencies
flask>=2.0.0
werkzeug>=2.0.0
moviepy>=2.0.0
pywebview>=4.0.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8
numpy>=1.20.0
